// src/shared/generated/MacroStepDefinition.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type MacroStepDefinition = { "type": "key", id: string, code: string, modifiers: Array<string>, action?: string, } | { "type": "click", id: string, unit?: string, anchor?: string, xPercent?: number, yPercent?: number, xPx?: number, yPx?: number, } | { "type": "delay", id: string, ms: number, } | { "type": "macro", id: string, macroId: string, callMode?: string, };